+++ intelmq/lib/upgrades.py
@@ -25,6 +25,7 @@ __all__ = ['v100_dev7_modify_syntax',
            'v220_configuration',
            'v220_azure_collector',
            'v220_feed_changes',
+           'v221_feed_changes',
            ]
 
 
@@ -117,7 +118,7 @@ def v110_shadowserver_feednames(defaults, runtime, harmonization, dry_run):
 
 def v110_deprecations(defaults, runtime, harmonization, dry_run):
     """
-    Checking for deprecated runtime configurations (stomp collector, cymru parser, ripe expert)
+    Checking for deprecated runtime configurations (stomp collector, cymru parser, ripe expert, collector feed parameter)
     """
     mapping = {
         "intelmq.bots.collectors.n6.collector_stomp": "intelmq.bots.collectors.stomp.collector",
@@ -462,6 +463,43 @@ def v220_feed_changes(defaults, runtime, harmonization, dry_run):
     return messages + ' Remove affected bots yourself.' if messages else changed, defaults, runtime, harmonization
 
 
+def v221_feed_changes(defaults, runtime, harmonization, dry_run):
+    """
+    Migrates feeds' configuration for changed/fixed parameters. Deprecation of HP Hosts file feed & parser.
+    """
+    found_hphosts_collector = []
+    found_hphosts_parser = []
+    messages = []
+    ULRHAUS_OLD = ['time.source', 'source.url', 'status', 'extra.urlhaus.threat_type', 'source.fqdn', 'source.ip', 'source.asn', 'source.geolocation.cc']
+    URLHAUS_NEW = ['time.source', 'source.url', 'status', 'classification.type|__IGNORE__', 'source.fqdn|__IGNORE__', 'source.ip', 'source.asn', 'source.geolocation.cc']
+    changed = None
+    for bot_id, bot in runtime.items():
+        if bot["module"] == "intelmq.bots.collectors.http.collector_http":
+            if bot["parameters"].get("http_url", None) == "http://hosts-file.net/download/hosts.txt":
+                found_hphosts_collector.append(bot_id)
+        elif bot['module'] == "intelmq.bots.parsers.hphosts.parser":
+            found_hphosts_parser.append(bot_id)
+        if bot["module"] == "intelmq.bots.parsers.generic.parser_csv":
+            if "columns" not in bot["parameters"]:
+                continue
+            columns = bot["parameters"]["columns"]
+            # convert columns to an array
+            if type(columns) is str:
+                columns = [column.strip() for column in columns.split(",")]
+            if columns == ULRHAUS_OLD:
+                changed = True
+                bot["parameters"]["columns"] = URLHAUS_NEW
+
+    if found_hphosts_collector:
+        messages.append('A discontinued feed "HP Hosts File" has been found '
+                        'as bot %s.' % ', '.join(sorted(found_hphosts_collector)))
+    if found_hphosts_parser:
+        messages.append('The removed parser "HP Hosts" has been found '
+                        'as bot %s.' % ', '.join(sorted(found_hphosts_parser)))
+    messages = ' '.join(messages)
+    return messages + ' Remove affected bots yourself.' if messages else changed, defaults, runtime, harmonization
+
+
 UPGRADES = OrderedDict([
     ((1, 0, 0, 'dev7'), (v100_dev7_modify_syntax, )),
     ((1, 1, 0), (v110_shadowserver_feednames, v110_deprecations)),
@@ -475,8 +513,8 @@ UPGRADES = OrderedDict([
     ((2, 1, 1), ()),
     ((2, 1, 2), ()),
     ((2, 1, 3), (v213_deprecations, v213_feed_changes)),
-    ((2, 1, 4), ()),
     ((2, 2, 0), (v220_configuration, v220_azure_collector, v220_feed_changes)),
+    ((2, 2, 1), (v221_feed_changes, )),
 ])
 
 ALWAYS = (harmonization, )