+++ intelmq/tests/bots/experts/cymru_whois/test_expert.py
@@ -32,22 +32,8 @@ EXAMPLE_OUTPUT6 = {"__type": "Event",
                    "destination.asn": 16876,
                    "destination.network": "2001:500:88::/48",
                    }
-UNICODE_INPUT = {"__type": "Event",
-                 "destination.ip": "200.236.128.1",
-                 "time.observation": "2015-01-01T00:00:00+00:00",
-                 }
-UNICODE_OUTPUT = {"__type": "Event",
-                  "destination.ip": "200.236.128.1",
-                  "time.observation": "2015-01-01T00:00:00+00:00",
-                  "destination.registry": "LACNIC",
-                  "destination.allocated": "2000-02-15T00:00:00+00:00",
-                  "destination.as_name": "Fundação de Desenvolvimento da Pesquisa, BR",
-                  "destination.geolocation.cc": "BR",
-                  "destination.asn": 10417,
-                  "destination.network": "200.236.128.0/18",
-                  }
 EMPTY_INPUT = {"__type": "Event",
-               "source.ip": "198.105.125.77",  # no result
+               "source.ip": "127.0.0.1",  # no result
                "time.observation": "2015-01-01T00:00:00+00:00",
                }
 EXAMPLE_6TO4_INPUT = {"__type": "Event",
@@ -103,11 +89,6 @@ class TestCymruExpertBot(test.BotTestCase, unittest.TestCase):
         self.input_message = EXAMPLE_INPUT6
         self.run_bot()
         self.assertMessageEqual(0, EXAMPLE_OUTPUT6)
-
-    def test_unicode_as_name(self):
-        self.input_message = UNICODE_INPUT
-        self.run_bot()
-        self.assertMessageEqual(0, UNICODE_OUTPUT)
 
     def test_empty_result(self):
         self.input_message = EMPTY_INPUT