+++ intelmq/tests/lib/test_upgrades.py
@@ -345,6 +345,75 @@ V220_FEED = {
     "module": "intelmq.bots.parsers.urlvir.parser",
 },
 }
+V221_FEED = {
+"abusech-urlhaus-columns-string-parser": {
+    "parameters": {
+        "column_regex_search": {},
+        "columns": "time.source,source.url,status,extra.urlhaus.threat_type,source.fqdn,source.ip,source.asn,source.geolocation.cc",
+        "default_url_protocol": "http://",
+        "delimiter": ",",
+        "filter_text": None,
+        "filter_type": None,
+        "skip_header": False,
+        "time_format": None,
+        "type": "c2server",
+        "type_translation": {
+            "malware_download": "malware-distribution"
+        }
+    },
+    "module": "intelmq.bots.parsers.generic.parser_csv",
+},
+"abusech-urlhaus-columns-dict-parser": {
+    "parameters": {
+        "column_regex_search": {},
+        "columns": ["time.source", "source.url","status","extra.urlhaus.threat_type","source.fqdn","source.ip","source.asn","source.geolocation.cc"],
+        "default_url_protocol": "http://",
+        "delimiter": ",",
+        "filter_text": None,
+        "filter_type": None,
+        "skip_header": False,
+        "time_format": None,
+        "type": "c2server",
+        "type_translation": {
+            "malware_download": "malware-distribution"
+        }
+    },
+    "module": "intelmq.bots.parsers.generic.parser_csv",
+}
+}
+V221_FEED_OUT = {
+"abusech-urlhaus-columns-string-parser": {
+    "parameters": {
+        "column_regex_search": {},
+        "columns": ['time.source', 'source.url', 'status', 'classification.type|__IGNORE__', 'source.fqdn|__IGNORE__', 'source.ip', 'source.asn', 'source.geolocation.cc'],
+        "default_url_protocol": "http://",
+        "delimiter": ",",
+        "filter_text": None,
+        "filter_type": None,
+        "skip_header": False,
+        "time_format": None,
+        "type": "c2server",
+        "type_translation": {
+            "malware_download": "malware-distribution"
+        }
+    },
+    "module": "intelmq.bots.parsers.generic.parser_csv",
+}
+}
+V221_FEED_OUT['abusech-urlhaus-columns-dict-parser'] = V221_FEED_OUT['abusech-urlhaus-columns-string-parser']
+V221_FEED_2 = {
+"hphosts-collector": {
+    "group": "Collector",
+    "module": "intelmq.bots.collectors.http.collector_http",
+    "parameters": {
+        "http_url": "http://hosts-file.net/download/hosts.txt",
+    },
+},
+"hphosts-parser": {
+    "group": "Parser",
+    "module": "intelmq.bots.parsers.hphosts.parser",
+},
+}
 
 
 def generate_function(function):
@@ -471,10 +540,28 @@ class TestUpgradeLib(unittest.TestCase):
                          result[0])
         self.assertEqual(V220_FEED, result[2])
 
+    def test_v221_feed_changes(self):
+        """ Test v221_feeds_1 """
+        result = upgrades.v221_feed_changes({}, V221_FEED, {}, False)
+        self.assertTrue(result[0])
+        self.assertEqual(V221_FEED_OUT, result[2])
+
+    def test_v221_feed_changes_2(self):
+        """ Test v213_feed_changes """
+        result = upgrades.v221_feed_changes({}, V221_FEED_2, {}, False)
+        self.assertEqual('A discontinued feed "HP Hosts File" has been found '
+                         'as bot hphosts-collector. '
+                         'The removed parser "HP Hosts" has been found '
+                         'as bot hphosts-parser. '
+                         'Remove affected bots yourself.',
+                         result[0])
+        self.assertEqual(V221_FEED_2, result[2])
+
 
 for name in upgrades.__all__:
     setattr(TestUpgradeLib, 'test_function_%s' % name,
             generate_function(getattr(upgrades, name)))
+
 
 if __name__ == '__main__':  # pragma: no cover
     unittest.main()